[Update] allow decimal rating

+++ SAE_S3.01/src/Entity/Rating.php
@@ -69,12 +69,12 @@ class Rating
         return $this->id;
     }
 
-    public function getValue(): ?int
+    public function getValue(): ?float
     {
         return $this->value;
     }
 
-    public function setValue(int $value): self
+    public function setValue(float $value): self
     {
         $this->value = $value;
